+++ packages/retail-ui/components/CustomComboBox/ComboBoxView.tsx
@@ -5,13 +5,13 @@ import DropdownContainer from '../DropdownContainer/DropdownContainer';
 import Input from '../Input';
 import InputLikeText from '../internal/InputLikeText';
 import Menu from '../Menu/Menu';
-import MenuItem, { MenuItemState } from '../MenuItem';
+import { MenuItemState } from '../MenuItem';
 import RenderLayer from '../RenderLayer';
 import Spinner from '../Spinner';
 import { Nullable } from '../../typings/utility-types';
 import { ArrowTriangleDown } from '@skbkontur/react-icons';
-
 import styles from './CustomComboBox.less';
+import ComboBoxMenu from './ComboBoxMenu';
 
 interface ComboBoxViewProps<T> {
   align?: 'left' | 'center' | 'right';
@@ -99,12 +99,17 @@ class ComboBoxView<T> extends React.Component<ComboBoxViewProps<T>> {
       onMouseLeave,
       onMouseOver,
       opened,
+      refMenu,
+      maxMenuHeight,
+      renderTotalCount,
+      renderItem,
+      renderNotFound,
+      totalCount,
       size,
       width
     } = this.props;
 
     const input = this.renderInput();
-    const menu = this.renderMenu();
 
     const topOffsets = {
       spinner: 6,
@@ -141,92 +146,24 @@ class ComboBoxView<T> extends React.Component<ComboBoxViewProps<T>> {
               offsetY={1}
               disablePortal={this.props.disablePortal}
             >
-              {menu}
+              <ComboBoxMenu
+                items={items}
+                loading={loading}
+                maxMenuHeight={maxMenuHeight}
+                onChange={this.handleItemSelect}
+                opened={opened}
+                refMenu={refMenu}
+                renderTotalCount={renderTotalCount}
+                renderItem={renderItem!}
+                renderNotFound={renderNotFound}
+                totalCount={totalCount}
+              />
             </DropdownContainer>
           )}
         </label>
       </RenderLayer>
     );
   }
-
-  private renderMenu(): React.ReactNode {
-    const {
-      opened,
-      items,
-      totalCount,
-      loading,
-      refMenu,
-      renderNotFound,
-      renderTotalCount,
-      maxMenuHeight
-    } = this.props;
-
-    if (!opened) {
-      return null;
-    }
-
-    if (loading && (!items || !items.length)) {
-      return (
-        <Menu ref={refMenu}>
-          <MenuItem disabled>
-            <div className={styles.menuItemSpinnerWrapper}>
-              <Spinner type="mini" dimmed />
-            </div>
-          </MenuItem>
-        </Menu>
-      );
-    }
-
-    if ((items == null || items.length === 0) && renderNotFound) {
-      return (
-        <Menu ref={refMenu}>
-          <MenuItem disabled>{renderNotFound()}</MenuItem>
-        </Menu>
-      );
-    }
-
-    let total = null;
-    if (items && renderTotalCount && totalCount && items.length < totalCount) {
-      total = (
-        <MenuItem disabled>
-          <div className={styles.menuItemContent}>
-            {renderTotalCount(items.length, totalCount)}
-          </div>
-        </MenuItem>
-      );
-    }
-
-    return (
-      <Menu ref={refMenu} maxHeight={maxMenuHeight}>
-        {items && items.map(this.renderItem)}
-        {total}
-      </Menu>
-    );
-  }
-
-  private renderItem = (item: T, index: number): React.ReactNode => {
-    // NOTE this is undesireable feature, better
-    // to remove it from further versions
-    if (typeof item === 'function' || React.isValidElement(item)) {
-      // @ts-ignore
-      const element = typeof item === 'function' ? item() : item;
-      const props = Object.assign(
-        {
-          key: index,
-          onClick: (e: React.SyntheticEvent) =>
-            this.handleItemSelect(element.props, e)
-        },
-        element.props
-      );
-      return React.cloneElement(element, props);
-    }
-    return (
-      // tslint:disable-next-line:jsx-no-lambda
-      <MenuItem onClick={e => this.handleItemSelect(item, e)} key={index}>
-        {state => this.props.renderItem!(item, state)}
-      </MenuItem>
-    );
-  };
 
   private renderInput(): React.ReactNode {
     const {