+++ packages/retail-ui/components/Select/__stories__/Select.stories.tsx
@@ -2,6 +2,8 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
 import Select from '../Select';
+import AddIcon from '@skbkontur/react-icons/Add';
+import { action } from '@storybook/addon-actions';
 
 class SelectWrapper extends React.Component<{}, any> {
   public state = {
@@ -72,8 +74,40 @@ storiesOf('Select', module)
   .add('With null', () => <SelectWithNull />)
   .add('use link', () => <Select use="link" items={['one', 'two', 'three']} />)
   .add('use link with icon', () => (
-    <Select _icon="Add" use="link" items={['one', 'two', 'three']} />
+    <Select _icon={<AddIcon />} use="link" items={['one', 'two', 'three']} />
   ))
   .add('with text overflow', () => (
     <Select width="100px" items={['oneoneone', 'twotwotwo', 'twotwotwo']} />
-  ));
+  ))
+  .add('external focus', () => {
+    class Sample extends React.Component {
+      private selectElem: Select | null = null;
+      public render() {
+        return (
+          <div>
+            <Select
+              width="100px"
+              items={['oneoneone', 'twotwotwo', 'twotwotwo']}
+              ref={this.refSelect}
+              onFocus={action('handleFocus')}
+              onBlur={action('handleBlur')}
+            />
+            <br />
+            <button onClick={this.handleClick}>Focus!</button>
+          </div>
+        );
+      }
+
+      private refSelect = (element: Select<any, any> | null) => {
+        this.selectElem = element;
+      };
+
+      private handleClick = () => {
+        if (this.selectElem) {
+          this.selectElem.focus();
+        }
+      };
+    }
+
+    return <Sample />;
+  });